Comments, fixes and refactoring

+++ targets/Main.cpp
@@ -8,8 +8,8 @@ int main() {
 	arch::Unique<arch::Application> application = arch::createUnique<vs::VulkanVs>();
 
 	auto engineConfig = arch::EngineConfig{
-		.windowWidth = vs::windowWidth,
-		.windowHeight = vs::windowHeight,
+		.windowWidth = (int)vs::windowWidth,
+		.windowHeight = (int)vs::windowHeight,
 		.windowTitle = "VulkanVs",
 		.backgroundColor = arch::Color(0, 0, 0, 0),
 		.renderingApi = arch::gfx::RenderingAPI::Nvrhi_VK
